started migration

+++ migrations/001.do.create-bookmarks.sql
@@ -3,5 +3,5 @@ CREATE TABLE bookmarks_list (
     title TEXT NOT NULL,
     url TEXT NOT NULL,
     description TEXT,
-    rating decimal(2,1)  NOT NULL
+    rating INTEGER  NOT NULL
 );
